src/requirements.txt
streamlit>=1.28.0
langchain>=0.1.0
langchain-community>=0.0.10
langchain-core>=0.1.0
chromadb>=0.4.0
sentence-transformers>=2.2.0
PyMuPDF>=1.23.0
pdfminer.six>=20221105
PyPDF2>=3.0.0
torch>=2.0.0
transformers>=4.30.0
numpy>=1.24.0
pandas>=2.0.0 